Diagram of ACT work
This diagram shows the hierarchy and organization of the ACT work, as well as who creates or benefits from each aspect of the work.  The ACT Rule format 1.0 is the bottom block, a specification that is the basis of the rules that will be developed.  This spec is written by the ACT Task Force and provides a common way to express plain English rules. The next level shows how the rules can be developed by various sources to become part of the ACT Rules repository.  The Auto-WCAG Community Group is one of the main groups currently developing rules in the ACT Rule Format, and there can be multiple sources or contributors of rules.  The ACT Rules will be collected in a rules repository after vetting through using the ACT Review Process to ensure the rules that are contributed are vetted for accuracy and the correct interpretation of the WCAG requirements.  There will be multiple implementations by the various test methodology and tool developers with the end result being more consistent test results for evaluators of accessible technology.