Unnecessary autocomplete support item (#2284)
Noticed this one reviewing our support items. This looks like it might have been added by some misunderstanding on our part, not something that was a problem in browsers.

The issue here related to form elements with autocomplete=off, but as far as I can tell that does not result in `elm.autocomplete === ''` when the autocomplete is otherwise set.